RavenDB-17364 - changed so database record is not fetched via route and added database admin check
